fix(api): change songs.global_seed from INTEGER to BIGINT
Fixed database schema issue where global_seed column was too small to handle timestamp-based seed values.

Changes:
- Updated Song model: global_seed from Integer to BigInteger
- Created Alembic migration (698242fd277f) to ALTER COLUMN type
- Supports values up to 2^63-1 instead of 2^31-1

Resolves: "integer out of range" error when creating songs with large seed values (e.g., 1763319120112)

services/api/app/models/song.py:51
services/api/alembic/versions/20251117_1430_698242fd277f_fix_songs_global_seed_bigint.py

+++ services/api/app/models/song.py
@@ -4,7 +4,7 @@ This module defines the Song and WorkflowRun SQLAlchemy ORM models that
 capture song metadata, SDS version, global seed, and workflow execution state.
 """
 
-from sqlalchemy import Column, String, Integer, Index, ForeignKey, CheckConstraint
+from sqlalchemy import Column, String, Integer, BigInteger, Index, ForeignKey, CheckConstraint
 from sqlalchemy.dialects.postgresql import UUID as SA_UUID, JSONB
 from sqlalchemy.orm import relationship
 
@@ -48,7 +48,7 @@ class Song(BaseModel):
 
     # Determinism seed
     global_seed = Column(
-        Integer,
+        BigInteger,
         nullable=False,
         comment="Global seed for deterministic generation (non-negative integer)"
     )
